+++ packages/starknet-snap/src/utils/constants.ts
@@ -25,11 +25,29 @@ export const STARKNET_MAINNET_NETWORK: Network = {
 };
 
 export const STARKNET_TESTNET_NETWORK: Network = {
-  name: 'Goerli Testnet',
+  name: 'Goerli Testnet (deprecated soon)',
   chainId: constants.StarknetChainId.SN_GOERLI,
   baseUrl: 'https://alpha4.starknet.io',
   nodeUrl: 'https://starknet-goerli.infura.io/v3/60c7253fb48147658095fe0460ac9ee9',
   voyagerUrl: 'https://goerli.voyager.online',
+  accountClassHash: '', // from argent-x repo
+};
+
+// TODO: Importing directly from constants when upgrading to starknet.js v6
+const SN_SEPOLIA = {
+  name: 'Sepolia Testnet',
+  baseUrl: 'https://alpha-sepolia.starknet.io',
+  chainId: '0x534e5f5345504f4c4941',
+  nodeUrl: 'https://starknet-sepolia.infura.io/v3/60c7253fb48147658095fe0460ac9ee9',
+  voyagerUrl: 'https://sepolia.voyager.online',
+};
+
+export const STARKNET_SEPOLIA_TESTNET_NETWORK: Network = {
+  name: SN_SEPOLIA.name,
+  chainId: SN_SEPOLIA.chainId,
+  baseUrl: SN_SEPOLIA.baseUrl,
+  nodeUrl: SN_SEPOLIA.nodeUrl,
+  voyagerUrl: SN_SEPOLIA.voyagerUrl,
   accountClassHash: '', // from argent-x repo
 };
 
@@ -58,6 +76,14 @@ export const ETHER_TESTNET: Erc20Token = {
   chainId: STARKNET_TESTNET_NETWORK.chainId,
 };
 
+export const ETHER_SEPOLIA_TESTNET: Erc20Token = {
+  address: '0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7',
+  name: 'Ether',
+  symbol: 'ETH',
+  decimals: 18,
+  chainId: SN_SEPOLIA.chainId,
+};
+
 export const DAI_MAINNET: Erc20Token = {
   address: '0x00da114221cb83fa859dbdb4c44beeaa0bb37c7537ad5ae66fe5e0efd20e6eb3',
   name: 'DAI',
@@ -72,6 +98,14 @@ export const DAI_TESTNET: Erc20Token = {
   symbol: 'DAI',
   decimals: 18,
   chainId: STARKNET_TESTNET_NETWORK.chainId,
+};
+
+export const DAI_SEPOLIA_TESTNET: Erc20Token = {
+  address: '0x03e85bfbb8e2a42b7bead9e88e9a1b19dbccf661471061807292120462396ec9',
+  name: 'DAI',
+  symbol: 'DAI',
+  decimals: 18,
+  chainId: SN_SEPOLIA.chainId,
 };
 
 export const WBTC_TESTNET: Erc20Token = {
@@ -98,6 +132,14 @@ export const USDC_TESTNET: Erc20Token = {
   chainId: STARKNET_TESTNET_NETWORK.chainId,
 };
 
+export const USDC_SEPOLIA_TESTNET: Erc20Token = {
+  address: '0x053c91253bc9682c04929ca02ed00b3e423f6710d2ee7e0d5ebb06f3ecf368a8',
+  name: 'USD Coin',
+  symbol: 'USDC',
+  decimals: 6,
+  chainId: SN_SEPOLIA.chainId,
+};
+
 export const USDT_MAINNET: Erc20Token = {
   address: '0x068f5c6a61780768455de69077e07e89787839bf8166decfbf92b645209c0fb8',
   name: 'Tether USD',
@@ -112,6 +154,14 @@ export const USDT_TESTNET: Erc20Token = {
   symbol: 'USDT',
   decimals: 6,
   chainId: STARKNET_TESTNET_NETWORK.chainId,
+};
+
+export const USDT_SEPOLIA_TESTNET: Erc20Token = {
+  address: '0x068f5c6a61780768455de69077e07e89787839bf8166decfbf92b645209c0fb8',
+  name: 'Tether USD',
+  symbol: 'USDT',
+  decimals: 6,
+  chainId: SN_SEPOLIA.chainId,
 };
 
 export const TEST_TOKEN_MAINNET: Erc20Token = {
@@ -130,7 +180,22 @@ export const TEST_TOKEN_TESTNET: Erc20Token = {
   chainId: STARKNET_TESTNET_NETWORK.chainId,
 };
 
-export const PRELOADED_TOKENS = [ETHER_MAINNET, ETHER_TESTNET, USDC_MAINNET, USDC_TESTNET, USDT_MAINNET, USDT_TESTNET];
-export const PRELOADED_NETWORKS = [STARKNET_MAINNET_NETWORK, STARKNET_TESTNET_NETWORK, STARKNET_INTEGRATION_NETWORK];
+export const PRELOADED_TOKENS = [
+  ETHER_MAINNET,
+  ETHER_TESTNET,
+  ETHER_SEPOLIA_TESTNET,
+  USDC_MAINNET,
+  USDC_TESTNET,
+  USDC_SEPOLIA_TESTNET,
+  USDT_MAINNET,
+  USDT_TESTNET,
+  USDT_SEPOLIA_TESTNET,
+];
+export const PRELOADED_NETWORKS = [
+  STARKNET_MAINNET_NETWORK,
+  STARKNET_TESTNET_NETWORK,
+  STARKNET_SEPOLIA_TESTNET_NETWORK,
+  STARKNET_INTEGRATION_NETWORK,
+];
 
 export const PROXY_CONTRACT_HASH = '0x25ec026985a3bf9d0cc1fe17326b245dfdc3ff89b8fde106542a3ea56c5a918'; // from argent-x repo