+++ packages/wallet-ui/src/utils/utils.ts
@@ -4,6 +4,8 @@ import {
   DECIMALS_DISPLAYED_MAX_LENGTH,
   STARKNET_MAINNET_EXPLORER,
   STARKNET_TESTNET_EXPLORER,
+  STARKNET_SEPOLIA_TESTNET_EXPLORER,
+  SEPOLIA_CHAINID,
   TIMEOUT_DURATION,
 } from './constants';
 import { Erc20Token, Erc20TokenBalance } from 'types';
@@ -26,6 +28,9 @@ export const openExplorerTab = (
       break;
     case constants.StarknetChainId.TESTNET:
       explorerUrl = STARKNET_TESTNET_EXPLORER;
+      break;
+    case SEPOLIA_CHAINID:
+      explorerUrl = STARKNET_SEPOLIA_TESTNET_EXPLORER;
       break;
   }
   window.open(explorerUrl + type + '/' + address, '_blank')?.focus();